+++ js/setup.js
@@ -1,0 +1,10 @@
+export const FLOAT_NUMBERS = 5;
+export const TOTAL_ADVERTS = 10;
+export const TOTAL_AVATAR_URLS = 10;
+export const MAX_PRICE = 1000;
+export const MAX_ROOMS = 10;
+export const ROOMS_TO_GUESTS_RATIO = 2;
+export const MIN_LAT = 35.65;
+export const MAX_LAT = 35.7;
+export const MIN_LNG = 139.7;
+export const MAX_LNG = 139.8;
